Added method for initializing the sensor with a pre-initialized channel
Multiplexers will change the addressing scheme for the sensor so the
convenient init method with searches in the predefined range between
`0x36..=0x39` will not work.

I added a method which take an `rppal::i2c::I2c` channel and does the
necessary checks on that to initialize the sensor.

This also turned out to need some refactoring of the other methods to
avoid duplicating the initialization code of which it's preferable
to only have one place that's written.

I also added some new errors so that we don't leak errors from `rppal`
in the "expected" cases.